+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/guide/SlimefunGuide.java
@@ -4,6 +4,7 @@ import java.util.Optional;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
@@ -11,6 +12,7 @@ import org.bukkit.inventory.ItemStack;
 import io.github.thebusybiscuit.slimefun4.api.player.PlayerProfile;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.implementation.guide.SurvivalSlimefunGuide;
+
 import io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
@@ -66,16 +68,19 @@ public final class SlimefunGuide {
         }
     }
 
+    @ParametersAreNonnullByDefault
     private static void openMainMenuAsync(Player player, SlimefunGuideMode layout, int selectedPage) {
         if (!PlayerProfile.get(player, profile -> SlimefunPlugin.runSync(() -> openMainMenu(profile, layout, selectedPage)))) {
             SlimefunPlugin.getLocalization().sendMessage(player, "messages.opening-guide");
         }
     }
 
+    @ParametersAreNonnullByDefault
     public static void openMainMenu(PlayerProfile profile, SlimefunGuideMode layout, int selectedPage) {
         SlimefunPlugin.getRegistry().getGuideLayout(layout).openMainMenu(profile, selectedPage);
     }
 
+    @ParametersAreNonnullByDefault
     public static void openCategory(PlayerProfile profile, Category category, SlimefunGuideMode layout, int selectedPage) {
         if (category == null) {
             return;
@@ -84,6 +89,7 @@ public final class SlimefunGuide {
         SlimefunPlugin.getRegistry().getGuideLayout(layout).openCategory(profile, category, selectedPage);
     }
 
+    @ParametersAreNonnullByDefault
     public static void openSearch(PlayerProfile profile, String input, boolean survival, boolean addToHistory) {
         SlimefunGuideImplementation layout = SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideMode.SURVIVAL_MODE);
 
@@ -94,10 +100,12 @@ public final class SlimefunGuide {
         layout.openSearch(profile, input, addToHistory);
     }
 
+    @ParametersAreNonnullByDefault
     public static void displayItem(PlayerProfile profile, ItemStack item, boolean addToHistory) {
         SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideMode.SURVIVAL_MODE).displayItem(profile, item, 0, addToHistory);
     }
 
+    @ParametersAreNonnullByDefault
     public static void displayItem(PlayerProfile profile, SlimefunItem item, boolean addToHistory) {
         SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideMode.SURVIVAL_MODE).displayItem(profile, item, addToHistory);
     }
@@ -106,6 +114,12 @@ public final class SlimefunGuide {
         return SlimefunUtils.isItemSimilar(item, getItem(SlimefunGuideMode.SURVIVAL_MODE), true) || SlimefunUtils.isItemSimilar(item, getItem(SlimefunGuideMode.CHEAT_MODE), true);
     }
 
+    /**
+     * Get the default layout for the Slimefun guide.
+     * Currently this is only {@link SlimefunGuideMode#SURVIVAL_MODE}.
+     *
+     * @return The default {@link SlimefunGuideLayout}.
+     */
     @Nonnull
     public static SlimefunGuideMode getDefaultLayout() {
         return SlimefunGuideMode.SURVIVAL_MODE;